+++ Net_Core/Framework/Enumerations/FunctionSecurityTableNames.cs
@@ -1,0 +1,8 @@
+namespace GrowthWare.Framework.Enumerations;
+
+public class FunctionSecurityTableNames
+{
+    public const string DERIVED_ROLES = "DerivedRoles";
+    public const string ASSIGNED_ROLES = "AssignedRoles";
+    public const string ASSIGNED_GROUPS = "Groups";
+}