+++ Net_Core/Framework/Enumerations/FunctionSecurityTables.cs
@@ -1,0 +1,8 @@
+namespace GrowthWare.Framework.Enumerations;
+
+public enum FunctionSecurityTables
+{
+    DerivedRoles = 0,
+    AssignedRoles = 1,
+    AssignedGroups = 2,
+}